pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langchain-camb"
version = "0.1.0"
description = "LangChain integration for CAMB AI - multilingual audio and localization tools"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "CAMB AI", email = "support@camb.ai" }
]
keywords = [
    "langchain",
    "camb",
    "tts",
    "text-to-speech",
    "translation",
    "transcription",
    "voice-cloning",
    "audio",
    "localization",
    "multilingual"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "langchain-core>=0.3.0",
    "camb-sdk>=1.5.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/camb-ai/langchain-camb"
Documentation = "https://docs.camb.ai"
Repository = "https://github.com/camb-ai/langchain-camb"

[tool.hatch.build.targets.wheel]
packages = ["langchain_camb"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (require CAMB_API_KEY)",
]
